+++ http/src/request/channel/invite/get_invite.rs
@@ -24,7 +24,7 @@ struct GetInviteFields {
 /// use twilight_http::Client;
 ///
 /// # #[tokio::main]
-/// # async fn main() -> Result<(), Box<dyn std::error::Error + Send + Sync>> {
+/// # async fn main() -> Result<(), Box<dyn std::error::Error>> {
 /// let client = Client::new("my token");
 ///
 /// let invite = client