+++ http/src/request/channel/update_channel_permission.rs
@@ -18,7 +18,7 @@ use twilight_model::{
 /// use twilight_model::id::{ChannelId, RoleId};
 ///
 /// # #[tokio::main]
-/// # async fn main() -> Result<(), Box<dyn std::error::Error + Send + Sync>> {
+/// # async fn main() -> Result<(), Box<dyn std::error::Error>> {
 /// let client = Client::new("my token");
 ///
 /// let channel_id = ChannelId(123);