+++ http/src/request/guild/emoji/get_emojis.rs
@@ -17,7 +17,7 @@ use twilight_model::{guild::Emoji, id::GuildId};
 /// use twilight_model::id::GuildId;
 ///
 /// # #[tokio::main]
-/// # async fn main() -> Result<(), Box<dyn std::error::Error + Send + Sync>> {
+/// # async fn main() -> Result<(), Box<dyn std::error::Error>> {
 /// let client = Client::new("my token");
 ///
 /// let guild_id = GuildId(100);