+++ http/src/request/guild/member/search_guild_members.rs
@@ -94,7 +94,7 @@ struct SearchGuildMembersFields {
 /// use twilight_model::id::GuildId;
 ///
 /// # #[tokio::main]
-/// # async fn main() -> Result<(), Box<dyn std::error::Error + Send + Sync>> {
+/// # async fn main() -> Result<(), Box<dyn std::error::Error>> {
 /// let client = Client::new("my token");
 ///
 /// let guild_id = GuildId(100);